requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.13
email-validator==1.3.1
python-multipart==0.0.9
web3==6.20.1
py-solc-x==2.0.4
rq==1.15.0
redis==5.0.0
python-dotenv==1.0.1